Move Haversine.distance() to Coordinates.haversineDistance()
